apps/carousel-agent/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "carousel-agent"
version = "0.1.0"
description = "Health-news -> Instagram carousel pipeline for @amarolabs, built on the OpenAI Agents SDK."
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }
dependencies = [
    "click>=8.1",
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "python-dateutil>=2.9",
]

[project.optional-dependencies]
agents = [
    "openai>=1.40",
    "openai-agents>=0.1",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "mypy>=1.10",
    "types-PyYAML",
]

[project.scripts]
carousel-agent = "carousel_agent.cli:cli"

[tool.setuptools.packages.find]
include = ["carousel_agent*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
carousel_agent = ["*.yaml"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
no_implicit_reexport = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
